chore: before introducing a new task manager UI, rename the current one as Legacy
before starting the new implementation, rename the current one.
it'll be removed when the new one will be over

related to https://github.com/podman-desktop/podman-desktop/issues/8628
